Make `pallet-posts` compilable
