+++ pallets/posts/Cargo.toml
@@ -1,8 +1,8 @@
 [package]
 name = 'pallet-posts'
-version = '0.7.3'
+version = '0.1.5'
 authors = ['DappForce <dappforce@pm.me>']
-edition = '2018'
+edition = '2021'
 license = 'GPL-3.0-only'
 homepage = 'https://subsocial.network'
 repository = 'https://github.com/dappforce/dappforce-subsocial-node'
@@ -18,30 +18,32 @@ std = [
     'scale-info/std',
     'frame-support/std',
     'frame-system/std',
+    'pallet-timestamp/std',
     'sp-runtime/std',
     'sp-std/std',
     'df-traits/std',
     'pallet-permissions/std',
     'pallet-space-follows/std',
     'pallet-spaces/std',
-    'pallet-utils/std',
+    'pallet-parachain-utils/std',
 ]
 
 [dependencies]
-codec = { package = "parity-scale-codec", version = "2.0.0", default-features = false, features = ["derive"] }
-scale-info = { version = "1.0", default-features = false, features = ["derive"] }
-serde = { features = ['derive'], optional = true, version = '1.0.119' }
-impl-trait-for-tuples = '0.1.3'
+impl-trait-for-tuples = '0.2'
+codec = { package = "parity-scale-codec", version = "3.0.0", default-features = false, features = ["derive"] }
+scale-info = { version = "2.1.1", default-features = false, features = ["derive"] }
+serde = { features = ['derive'], optional = true, version = '1.0.132' }
 
 # Local dependencies
 df-traits = { default-features = false, path = '../traits' }
 pallet-permissions = { default-features = false, path = '../permissions' }
 pallet-space-follows = { default-features = false, path = '../space-follows' }
 pallet-spaces = { default-features = false, path = '../spaces' }
-pallet-utils = { default-features = false, path = '../utils' }
+pallet-parachain-utils = { default-features = false, path = '../parachain-utils' }
 
 # Substrate dependencies
-frame-support = { git = 'https://github.com/paritytech/substrate', branch = 'polkadot-v0.9.12', default-features = false }
-frame-system = { git = 'https://github.com/paritytech/substrate', branch = 'polkadot-v0.9.12', default-features = false }
-sp-runtime = { git = 'https://github.com/paritytech/substrate', branch = 'polkadot-v0.9.12', default-features = false }
-sp-std = { git = 'https://github.com/paritytech/substrate', branch = 'polkadot-v0.9.12', default-features = false }
+frame-support = { git = 'https://github.com/paritytech/substrate', branch = 'polkadot-v0.9.19', default-features = false }
+frame-system = { git = 'https://github.com/paritytech/substrate', branch = 'polkadot-v0.9.19', default-features = false }
+pallet-timestamp = { git = 'https://github.com/paritytech/substrate', branch = 'polkadot-v0.9.19', default-features = false }
+sp-runtime = { git = 'https://github.com/paritytech/substrate', branch = 'polkadot-v0.9.19', default-features = false }
+sp-std = { git = 'https://github.com/paritytech/substrate', branch = 'polkadot-v0.9.19', default-features = false }
