+++ pallets/posts/src/lib.rs
@@ -28,112 +28,38 @@ use sp_std::prelude::*;
 
 use df_traits::moderation::{IsAccountBlocked, IsContentBlocked, IsPostBlocked};
 use pallet_permissions::SpacePermission;
-use pallet_spaces::{Module as Spaces, Space, SpaceById};
-use pallet_utils::{Content, Error as UtilsError, Module as Utils, PostId, SpaceId, WhoAndWhen};
+use pallet_spaces::{Pallet as Spaces, types::Space, SpaceById};
+use pallet_parachain_utils::{
+    Content, Error as UtilsError, PostId, SpaceId, WhoAndWhenOf, new_who_and_when,
+    throw_utils_error, ensure_content_is_valid,
+};
 
 pub use pallet::*;
 pub mod functions;
 
-pub mod rpc;
+pub mod types;
+pub use types::*;
 
-/// Information about a post's owner, its' related space, content, and visibility.
-#[derive(Encode, Decode, Clone, Eq, PartialEq, RuntimeDebug, TypeInfo)]
-#[scale_info(skip_type_params(T))]
-pub struct Post<T: Config> {
-    /// Unique sequential identifier of a post. Examples of post ids: `1`, `2`, `3`, and so on.
-    pub id: PostId,
-
-    pub created: WhoAndWhen<T>,
-    pub updated: Option<WhoAndWhen<T>>,
-
-    /// The current owner of a given post.
-    pub owner: T::AccountId,
-
-    /// Through post extension you can provide specific information necessary for different kinds
-    /// of posts such as regular posts, comments, and shared posts.
-    pub extension: PostExtension,
-
-    /// An id of a space which contains a given post.
-    pub space_id: Option<SpaceId>,
-
-    pub content: Content,
-
-    /// Hidden field is used to recommend to end clients (web and mobile apps) that a particular
-    /// posts and its' comments should not be shown.
-    pub hidden: bool,
-
-    /// The total number of replies for a given post.
-    pub replies_count: u16,
-
-    /// The number of hidden replies for a given post.
-    pub hidden_replies_count: u16,
-
-    /// The number of times a given post has been shared.
-    pub shares_count: u16,
-
-    /// The number of times a given post has been upvoted.
-    pub upvotes_count: u16,
-
-    /// The number of times a given post has been downvoted.
-    pub downvotes_count: u16,
-
-    pub score: i32,
-}
-
-#[derive(Encode, Decode, Clone, Eq, PartialEq, RuntimeDebug, TypeInfo)]
-pub struct PostUpdate {
-    /// Deprecated: This field has no effect in `fn update_post()` extrinsic.
-    /// See `fn move_post()` extrinsic if you want to move a post to another space.
-    pub space_id: Option<SpaceId>,
-
-    pub content: Option<Content>,
-    pub hidden: Option<bool>,
-}
-
-/// Post extension provides specific information necessary for different kinds
-/// of posts such as regular posts, comments, and shared posts.
-#[derive(Encode, Decode, Clone, Copy, Eq, PartialEq, RuntimeDebug, TypeInfo)]
-#[cfg_attr(feature = "std", derive(Serialize, Deserialize))]
-#[cfg_attr(feature = "std", serde(untagged))]
-pub enum PostExtension {
-    RegularPost,
-    Comment(Comment),
-    SharedPost(PostId),
-}
-
-#[derive(Encode, Decode, Clone, Copy, Eq, PartialEq, RuntimeDebug, TypeInfo)]
-#[cfg_attr(feature = "std", derive(Serialize, Deserialize))]
-pub struct Comment {
-    pub parent_id: Option<PostId>,
-    pub root_post_id: PostId,
-}
-
-impl Default for PostExtension {
-    fn default() -> Self {
-        PostExtension::RegularPost
-    }
-}
+// pub mod rpc;
 
 #[impl_trait_for_tuples::impl_for_tuples(10)]
 pub trait AfterPostUpdated<T: Config> {
     fn after_post_updated(account: T::AccountId, post: &Post<T>, old_data: PostUpdate);
 }
 
-pub const FIRST_POST_ID: u64 = 1;
-
 #[frame_support::pallet]
 pub mod pallet {
     use super::*;
+
     use frame_support::pallet_prelude::*;
     use frame_support::traits::IsType;
     use frame_system::pallet_prelude::*;
 
     #[pallet::config]
-    pub trait Config:
-        frame_system::Config
-        + pallet_utils::Config
+    pub trait Config: frame_system::Config
         + pallet_space_follows::Config
         + pallet_spaces::Config
+        + pallet_timestamp::Config
     {
         /// The overarching event type.
         type Event: From<Event<Self>> + IsType<<Self as frame_system::Config>::Event>;
@@ -149,7 +75,8 @@ pub mod pallet {
 
     #[pallet::pallet]
     #[pallet::generate_store(pub(super) trait Store)]
-    pub struct Pallet<T>(PhantomData<T>);
+    #[pallet::without_storage_info]
+    pub struct Pallet<T>(_);
 
     #[pallet::hooks]
     impl<T: Config> Hooks<BlockNumberFor<T>> for Pallet<T> {
@@ -169,211 +96,7 @@ pub mod pallet {
         }
     }
 
-    #[pallet::call]
-    impl<T: Config> Pallet<T> {
-        #[pallet::weight(100_000 + T::DbWeight::get().reads_writes(8, 8))]
-        pub fn create_post(
-            origin: OriginFor<T>,
-            space_id_opt: Option<SpaceId>,
-            extension: PostExtension,
-            content: Content,
-        ) -> DispatchResult {
-            use frame_support::StorageMap;
-            let creator = ensure_signed(origin)?;
-
-            Utils::<T>::is_valid_content(content.clone())?;
-
-            let new_post_id = Self::next_post_id();
-            let new_post: Post<T> = Post::new(
-                new_post_id,
-                creator.clone(),
-                space_id_opt,
-                extension,
-                content.clone(),
-            );
-
-            // Get space from either space_id_opt or Comment if a comment provided
-            let space = &mut new_post.get_space()?;
-            ensure!(!space.hidden, Error::<T>::CannotCreateInHiddenScope);
-
-            ensure!(
-                T::IsAccountBlocked::is_allowed_account(creator.clone(), space.id),
-                UtilsError::<T>::AccountIsBlocked
-            );
-            ensure!(
-                T::IsContentBlocked::is_allowed_content(content, space.id),
-                UtilsError::<T>::ContentIsBlocked
-            );
-
-            let root_post = &mut new_post.get_root_post()?;
-            ensure!(!root_post.hidden, Error::<T>::CannotCreateInHiddenScope);
-
-            // Check whether account has permission to create Post (by extension)
-            let mut permission_to_check = SpacePermission::CreatePosts;
-            let mut error_on_permission_failed = Error::<T>::NoPermissionToCreatePosts;
-
-            if let PostExtension::Comment(_) = extension {
-                permission_to_check = SpacePermission::CreateComments;
-                error_on_permission_failed = Error::<T>::NoPermissionToCreateComments;
-            }
-
-            Spaces::ensure_account_has_space_permission(
-                creator.clone(),
-                &space,
-                permission_to_check,
-                error_on_permission_failed.into(),
-            )?;
-
-            match extension {
-                PostExtension::RegularPost => space.inc_posts(),
-                PostExtension::SharedPost(post_id) => {
-                    Self::create_sharing_post(&creator, new_post_id, post_id, space)?
-                }
-                PostExtension::Comment(comment_ext) => {
-                    Self::create_comment(new_post_id, comment_ext, root_post)?
-                }
-            }
-
-            if new_post.is_root_post() {
-                SpaceById::<T>::insert(space.id, space.clone());
-                PostIdsBySpaceId::<T>::mutate(space.id, |ids| ids.push(new_post_id));
-            }
-
-            PostById::insert(new_post_id, new_post);
-            NextPostId::<T>::mutate(|n| {
-                *n += 1;
-            });
-
-            Self::deposit_event(Event::PostCreated(creator, new_post_id));
-            Ok(())
-        }
-
-        #[pallet::weight(100_000 + T::DbWeight::get().reads_writes(5, 3))]
-        pub fn update_post(
-            origin: OriginFor<T>,
-            post_id: PostId,
-            update: PostUpdate,
-        ) -> DispatchResult {
-            use frame_support::StorageMap;
-            let editor = ensure_signed(origin)?;
-
-            let has_updates = update.content.is_some() || update.hidden.is_some();
-
-            ensure!(has_updates, Error::<T>::NoUpdatesForPost);
-
-            let mut post = Self::require_post(post_id)?;
-            let mut space_opt = post.try_get_space();
-
-            if let Some(space) = &space_opt {
-                ensure!(
-                    T::IsAccountBlocked::is_allowed_account(editor.clone(), space.id),
-                    UtilsError::<T>::AccountIsBlocked
-                );
-                Self::ensure_account_can_update_post(&editor, &post, space)?;
-            }
-
-            let mut is_update_applied = false;
-            let mut old_data = PostUpdate::default();
-
-            if let Some(content) = update.content {
-                if content != post.content {
-                    Utils::<T>::is_valid_content(content.clone())?;
-
-                    if let Some(space) = &space_opt {
-                        ensure!(
-                            T::IsContentBlocked::is_allowed_content(content.clone(), space.id),
-                            UtilsError::<T>::ContentIsBlocked
-                        );
-                    }
-
-                    old_data.content = Some(post.content.clone());
-                    post.content = content;
-                    is_update_applied = true;
-                }
-            }
-
-            if let Some(hidden) = update.hidden {
-                if hidden != post.hidden {
-                    space_opt = space_opt.map(|mut space| {
-                        if hidden {
-                            space.inc_hidden_posts();
-                        } else {
-                            space.dec_hidden_posts();
-                        }
-
-                        space
-                    });
-
-                    if let PostExtension::Comment(comment_ext) = post.extension {
-                        Self::update_counters_on_comment_hidden_change(&comment_ext, hidden)?;
-                    }
-
-                    old_data.hidden = Some(post.hidden);
-                    post.hidden = hidden;
-                    is_update_applied = true;
-                }
-            }
-
-            // Update this post only if at least one field should be updated:
-            if is_update_applied {
-                post.updated = Some(WhoAndWhen::<T>::new(editor.clone()));
-
-                if let Some(space) = space_opt {
-                    SpaceById::<T>::insert(space.id, space);
-                }
-
-                <PostById<T>>::insert(post.id, post.clone());
-                T::AfterPostUpdated::after_post_updated(editor.clone(), &post, old_data);
-
-                Self::deposit_event(Event::PostUpdated(editor, post_id));
-            }
-            Ok(())
-        }
-
-        #[pallet::weight(T::DbWeight::get().reads(1) + 50_000)]
-        pub fn move_post(
-            origin: OriginFor<T>,
-            post_id: PostId,
-            new_space_id: Option<SpaceId>,
-        ) -> DispatchResult {
-            let who = ensure_signed(origin)?;
-
-            let post = &mut Self::require_post(post_id)?;
-
-            ensure!(
-                new_space_id != post.space_id,
-                Error::<T>::CannotMoveToSameSpace
-            );
-
-            if let Some(space) = post.try_get_space() {
-                Self::ensure_account_can_update_post(&who, &post, &space)?;
-            } else {
-                post.ensure_owner(&who)?;
-            }
-
-            let old_space_id = post.space_id;
-
-            if let Some(space_id) = new_space_id {
-                Self::move_post_to_space(who.clone(), post, space_id)?;
-            } else {
-                Self::delete_post_from_space(post_id)?;
-            }
-
-            let historical_data = PostUpdate {
-                space_id: old_space_id,
-                content: None,
-                hidden: None,
-            };
-
-            T::AfterPostUpdated::after_post_updated(who.clone(), &post, historical_data);
-
-            Self::deposit_event(Event::PostMoved(who, post_id));
-            Ok(())
-        }
-    }
-
     #[pallet::event]
-    #[pallet::metadata(T::AccountId = "AccountId")]
     #[pallet::generate_deposit(pub(super) fn deposit_event)]
     pub enum Event<T: Config> {
         PostCreated(T::AccountId, PostId),
@@ -439,6 +162,7 @@ pub mod pallet {
         /// A comment owner is not allowed to update their own comments in this space.
         NoPermissionToUpdateOwnComments,
     }
+    
     #[pallet::type_value]
     pub fn DefaultForNextPostId() -> PostId {
         FIRST_POST_ID
@@ -471,4 +195,205 @@ pub mod pallet {
     #[pallet::getter(fn shared_post_ids_by_original_post_id)]
     pub type SharedPostIdsByOriginalPostId<T: Config> =
         StorageMap<_, Twox64Concat, PostId, Vec<PostId>, ValueQuery>;
+
+    #[pallet::call]
+    impl<T: Config> Pallet<T> {
+        #[pallet::weight(100_000 + T::DbWeight::get().reads_writes(8, 8))]
+        pub fn create_post(
+            origin: OriginFor<T>,
+            space_id_opt: Option<SpaceId>,
+            extension: PostExtension,
+            content: Content,
+        ) -> DispatchResult {
+            let creator = ensure_signed(origin)?;
+
+            ensure_content_is_valid(content.clone())?;
+
+            let new_post_id = Self::next_post_id();
+            let new_post: Post<T> = Post::new(
+                new_post_id,
+                creator.clone(),
+                space_id_opt,
+                extension,
+                content.clone(),
+            );
+
+            // Get space from either space_id_opt or Comment if a comment provided
+            let space = &mut new_post.get_space()?;
+            ensure!(!space.hidden, Error::<T>::CannotCreateInHiddenScope);
+
+            ensure!(
+                T::IsAccountBlocked::is_allowed_account(creator.clone(), space.id),
+                throw_utils_error(UtilsError::AccountIsBlocked)
+            );
+            ensure!(
+                T::IsContentBlocked::is_allowed_content(content, space.id),
+                throw_utils_error(UtilsError::ContentIsBlocked)
+            );
+
+            let root_post = &mut new_post.get_root_post()?;
+            ensure!(!root_post.hidden, Error::<T>::CannotCreateInHiddenScope);
+
+            // Check whether account has permission to create Post (by extension)
+            let mut permission_to_check = SpacePermission::CreatePosts;
+            let mut error_on_permission_failed = Error::<T>::NoPermissionToCreatePosts;
+
+            if let PostExtension::Comment(_) = extension {
+                permission_to_check = SpacePermission::CreateComments;
+                error_on_permission_failed = Error::<T>::NoPermissionToCreateComments;
+            }
+
+            Spaces::ensure_account_has_space_permission(
+                creator.clone(),
+                &space,
+                permission_to_check,
+                error_on_permission_failed.into(),
+            )?;
+
+            match extension {
+                PostExtension::RegularPost => space.inc_posts(),
+                PostExtension::SharedPost(post_id) => {
+                    Self::create_sharing_post(&creator, new_post_id, post_id, space)?
+                }
+                PostExtension::Comment(comment_ext) => {
+                    Self::create_comment(new_post_id, comment_ext, root_post)?
+                }
+            }
+
+            if new_post.is_root_post() {
+                SpaceById::<T>::insert(space.id, space.clone());
+                PostIdsBySpaceId::<T>::mutate(space.id, |ids| ids.push(new_post_id));
+            }
+
+            PostById::insert(new_post_id, new_post);
+            NextPostId::<T>::mutate(|n| {
+                *n += 1;
+            });
+
+            Self::deposit_event(Event::PostCreated(creator, new_post_id));
+            Ok(())
+        }
+
+        #[pallet::weight(100_000 + T::DbWeight::get().reads_writes(5, 3))]
+        pub fn update_post(
+            origin: OriginFor<T>,
+            post_id: PostId,
+            update: PostUpdate,
+        ) -> DispatchResult {
+            let editor = ensure_signed(origin)?;
+
+            let has_updates = update.content.is_some() || update.hidden.is_some();
+
+            ensure!(has_updates, Error::<T>::NoUpdatesForPost);
+
+            let mut post = Self::require_post(post_id)?;
+            let mut space_opt = post.try_get_space();
+
+            if let Some(space) = &space_opt {
+                ensure!(
+                    T::IsAccountBlocked::is_allowed_account(editor.clone(), space.id),
+                    throw_utils_error(UtilsError::AccountIsBlocked)
+                );
+                Self::ensure_account_can_update_post(&editor, &post, space)?;
+            }
+
+            let mut is_update_applied = false;
+            let mut old_data = PostUpdate::default();
+
+            if let Some(content) = update.content {
+                if content != post.content {
+                    ensure_content_is_valid(content.clone())?;
+
+                    if let Some(space) = &space_opt {
+                        ensure!(
+                            T::IsContentBlocked::is_allowed_content(content.clone(), space.id),
+                            throw_utils_error(UtilsError::ContentIsBlocked)
+                        );
+                    }
+
+                    old_data.content = Some(post.content.clone());
+                    post.content = content;
+                    is_update_applied = true;
+                }
+            }
+
+            if let Some(hidden) = update.hidden {
+                if hidden != post.hidden {
+                    space_opt = space_opt.map(|mut space| {
+                        if hidden {
+                            space.inc_hidden_posts();
+                        } else {
+                            space.dec_hidden_posts();
+                        }
+
+                        space
+                    });
+
+                    if let PostExtension::Comment(comment_ext) = post.extension {
+                        Self::update_counters_on_comment_hidden_change(&comment_ext, hidden)?;
+                    }
+
+                    old_data.hidden = Some(post.hidden);
+                    post.hidden = hidden;
+                    is_update_applied = true;
+                }
+            }
+
+            // Update this post only if at least one field should be updated:
+            if is_update_applied {
+                post.updated = Some(new_who_and_when::<T>(editor.clone()));
+
+                if let Some(space) = space_opt {
+                    SpaceById::<T>::insert(space.id, space);
+                }
+
+                <PostById<T>>::insert(post.id, post.clone());
+                T::AfterPostUpdated::after_post_updated(editor.clone(), &post, old_data);
+
+                Self::deposit_event(Event::PostUpdated(editor, post_id));
+            }
+            Ok(())
+        }
+
+        #[pallet::weight(T::DbWeight::get().reads(1) + 50_000)]
+        pub fn move_post(
+            origin: OriginFor<T>,
+            post_id: PostId,
+            new_space_id: Option<SpaceId>,
+        ) -> DispatchResult {
+            let who = ensure_signed(origin)?;
+
+            let post = &mut Self::require_post(post_id)?;
+
+            ensure!(
+                new_space_id != post.space_id,
+                Error::<T>::CannotMoveToSameSpace
+            );
+
+            if let Some(space) = post.try_get_space() {
+                Self::ensure_account_can_update_post(&who, &post, &space)?;
+            } else {
+                post.ensure_owner(&who)?;
+            }
+
+            let old_space_id = post.space_id;
+
+            if let Some(space_id) = new_space_id {
+                Self::move_post_to_space(who.clone(), post, space_id)?;
+            } else {
+                Self::delete_post_from_space(post_id)?;
+            }
+
+            let historical_data = PostUpdate {
+                space_id: old_space_id,
+                content: None,
+                hidden: None,
+            };
+
+            T::AfterPostUpdated::after_post_updated(who.clone(), &post, historical_data);
+
+            Self::deposit_event(Event::PostMoved(who, post_id));
+            Ok(())
+        }
+    }
 }
